Add eunit_formatters for better output
This provides the "series of dots" style output for the tests that is common in many other test frameworks. This is also the default output mode when rebar3 runs eunit tests. Together with the change to only emit a single status report at the end of the run, this provides very nice output.